all: Move SignalHandler into the driver
