--- agent/src/main/native/com/code_intelligence/jazzer/runtime/BUILD.bazel
@@ -1,8 +1,0 @@
-load("@fmeum_rules_jni//jni:defs.bzl", "cc_jni_library")
-
-cc_jni_library(
-    name = "jazzer_signal_handler",
-    srcs = ["signal_handler.cpp"],
-    visibility = ["//agent/src/main/java/com/code_intelligence/jazzer/runtime:__pkg__"],
-    deps = ["//agent/src/main/java/com/code_intelligence/jazzer/runtime:signal_handler.hdrs"],
-)
