+++ driver/src/main/java/com/code_intelligence/jazzer/driver/BUILD.bazel
@@ -29,6 +29,7 @@ java_jni_library(
     deps = [
         ":opt",
         ":reproducer_template",
+        ":signal_handler",
         ":utils",
         "//agent/src/main/java/com/code_intelligence/jazzer/api",
         "//agent/src/main/java/com/code_intelligence/jazzer/autofuzz",
@@ -36,7 +37,6 @@ java_jni_library(
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime:coverage_map",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime:fuzzed_data_provider",
-        "//agent/src/main/java/com/code_intelligence/jazzer/runtime:signal_handler",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime:unsafe_provider",
         "//agent/src/main/java/com/code_intelligence/jazzer/utils",
     ],
@@ -56,6 +56,13 @@ java_library(
         "//agent/src/main/java/com/code_intelligence/jazzer:__subpackages__",
         "//driver/src/test/java/com/code_intelligence/jazzer/driver:__pkg__",
     ],
+)
+
+java_jni_library(
+    name = "signal_handler",
+    srcs = ["SignalHandler.java"],
+    native_libs = ["//driver/src/main/native/com/code_intelligence/jazzer/driver:jazzer_signal_handler"],
+    visibility = ["//driver/src/main/native/com/code_intelligence/jazzer/driver:__pkg__"],
 )
 
 java_library(
