+++ driver/src/main/java/com/code_intelligence/jazzer/driver/SignalHandler.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.code_intelligence.jazzer.runtime;
+package com.code_intelligence.jazzer.driver;
 
 import com.github.fmeum.rules_jni.RulesJni;
 import sun.misc.Signal;
